create placement API wsgi entry point
We want a wsgi_script as the entry point for our placement API
actually getting run, this is wrapping in the smooth and mellow pbr
patterns that make it sensible to also run just by starting it on the
command line.

This also actually initializes the logging subsystem for the placement
API, and does the standard pattern of dumping the configuration if
DEBUG is enabled. Pieces of this were cribbed/inspired by equivalent
keystone code.

The config directory is now setable via environment, which may be be
needed by folks with venvs.

Change-Id: I00d032554de273d7493cfb467f81687c08fd5389