[placement] Allow inventory to violate allocations
When a compute node is reconfigured to have different inventory it
needs to be able to tell the placement service of these changes,
even if they violate existing allocations. Thus, when updating
inventory we no longer check the new capacity against existing
allocations, we just let it happen (but log a warning).

This is safe because allocations in excess of capacity will still
prevent resources from being used.

Change-Id: I48ae1d7cd6bc309243493ddac99ce990c0146534
Closes-Bug: #1619722
Co-Authored-By: Dan Smith <dansmith@redhat.com>