[placement] annotate loadapp as public interface
When Change-Id: I61d20c5d19797f7e66648c7864a632f3328be8ce was under
review there were questions about why the interface on deploy() was not
changed to remove the no-longer used project_name kwarg. I mistakenly
asserted that this was because it was a public interface for building
a Placement WSGI application. That's not the case. loadapp() is the
public interface.

This change annotates loadapp() as such, and removes the unused arg
deploy().

Change-Id: I9f3d03d964654ab1b9515ecd6fe35c518d5f496a
Related-Bug: #1734491