Isolate placement database config
The placement database connection (which can use the same connection
string as the api database) needs to be managed from its own module so
that the nova db files are not imported as this eventually leads to
importing object-related and other code which the placement service does
not need.

The wsgi application is now responsible for initializing the database
configuration.

Fixtures and a small number of tests are updated to reflected the new
location of the placement engine.

The original parse_args needed to import RPC and database related
modules that are not required in the placement context. A new
_parse_args method is added to wsgi.py which does the minimal required
work.

blueprint placement-extract

Change-Id: I2fff528060ec52a4a2e26a6484bdf18359b95f77