Enable unit test discovery
This very large change represents what was needed to get to tox -epy27
to successfully discover the unit tests in placement/tests/unit

The tests then run but fail in four important ways:

* the resource provider object tests are turned off entirely as they
  use functionality in test_objects which though we can import fails
  upon trying to use entry points from nova (see next point)
* the base test in nova/test.py is replaces by a testtools.TestCase, as
  a simple standin. This avoids the entry point problem mentioned above
  but leaves us without some fixtures including:
* Log handling is not set up, so log trapping, provided by the nova.test
  base is not there
* self.flags is not present on the test case

These things will need to be fixed. Functional tests require much more
complex fixtures (for database handling) so that will be an important
next milestone to deal with.