Correct test-requirements.txt and stop test_report_client
test-requirements.txt got clobbered and was not reflecting the real
requirements.

test_report_client tests nova code, so should not be here, but I don't
want to remove it just yet.

At this stage the functional tests want to run but there are dependent
on many nova fixturs, and of course on the database schema. Until we
have those things, not much going on here.